+++ Renderer/RendererBackend/OpenGLRenderer/src/Texture/Texture2DBind.cpp
@@ -134,7 +134,7 @@ namespace OpenGLRenderer
 
 		#ifndef OPENGLRENDERER_NO_STATE_CLEANUP
 			// Be polite and restore the previous bound OpenGL texture
-			glBindTexture(GL_TEXTURE_2D, openGLTextureBackup);
+			glBindTexture(GL_TEXTURE_2D, static_cast<GLuint>(openGLTextureBackup));
 
 			// Restore previous alignment
 			glPixelStorei(GL_UNPACK_ALIGNMENT, openGLAlignmentBackup);
